Disable WP Rocket's page caching
WP Rocket's page caching and the Dynamic cache of SiteGround's SuperCacher plugin are currently not playing nice with each other. From SiteGround support: 

"The problem itself occurs due to the fact that WP Rocket's full page caching option caches the response headers returned by the web server when serving requests for a website. More specifically, it also caches a specific HTTP response header that we implement to determine whether the content of a page was cached by the SuperCacher. This is the X-Proxy-Cache header. [...] For the time being I recommend disabling full page caching with WP Rocket and relying on the SuperCacher dynamic cache instead. [...] With just full page caching disabled with WP Rocket you should still be able to benefit from all the rest of the plugin's features, which should be working in conjunction with the SuperCacher without any troubles."